# Fix: Ignore computed fields during sampling key extraction