--- a/config/config.go
+++ b/config/config.go
@@ -242,6 +242,7 @@ const RootPrefix = "root."
 // GetKeyFields returns the fields that should be used as keys for the sampler.
 // It returns two slices: the first contains all fields, including those with the root prefix,
 // and the second contains fields that do not have the root prefix.
+// Fields that start with "computed." are ignored since they should not exist as a field in a trace.
 func GetKeyFields(fields []string) (allFields []string, nonRootFields []string) {
 	if len(fields) == 0 {
 		return nil, nil
@@ -251,11 +252,26 @@ func GetKeyFields(fields []string) (allFields []string, nonRootFields []string) 
 	nonRootFields = make([]string, 0, len(fields))
 
 	for _, field := range fields {
-		if strings.HasPrefix(field, RootPrefix) {
-			rootFields = append(rootFields, field[len(RootPrefix):])
-		} else {
+		values := strings.SplitAfterN(field, ".", 2)
+		if len(values) < 2 {
+			nonRootFields = append(nonRootFields, field)
+			continue
+		}
+
+		switch values[0] {
+		case RootPrefix:
+			// If the field starts with "root.", add it to rootFields
+			rootFields = append(rootFields, values[1])
+		case ComputedFieldPrefix:
+			// If the field starts with "computed.", skip it
+		default:
+			// Otherwise, add it to nonRootFields
 			nonRootFields = append(nonRootFields, field)
 		}
+	}
+
+	if len(rootFields) == 0 && len(nonRootFields) == 0 {
+		return nil, nil
 	}
 
 	if len(rootFields) == 0 {