## Which problem is this PR solving?

Refinery was attempting to extract computed fields (prefixed with `?.`) like `?.NUMBER_DESCENDANTS` as if they were regular trace fields during the unmarshaling process. This caused unnecessary performance overhead since computed fields are virtual fields calculated during rule evaluation and don't exist as actual fields in trace data.

## Short description of the changes

- Ignore computed fields in `GetKeyFields` function