@VinozzZ 2025-07-29
> Basic concept is sound, but I think it's worth a quick performance check.
> 
> Gah, my comment got lost! I don't think the `SplitAfterN` function is going to be anywhere near as fast as a simple HasPrefix in the switch statement. I think you should benchmark it to make sure.
> 
> ```
> 			switch {
> 			case strings.HasPrefix(key, RootPrefix):
> ```

I tried out the string.HasPrefix version. The benchmark is surprising that the `SplitAfterN` is faster
```
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/refinery/types
cpu: Apple M2 Max
                                             │   split-after.txt   │              has-prefix.txt              │
                                             │   sec/op    │   sec/op     vs base                │
UnmarshalPayload/UnmarshalPayload-12           16.25µ ± 1%   17.82µ ± 1%   +9.67% (p=0.000 n=10)
UnmarshalPayload/UnmarshalPayloadComplete-12   15.60µ ± 1%   17.43µ ± 1%  +11.77% (p=0.000 n=10)
geomean                                        15.92µ        17.63µ       +10.72%

                                             │   new.txt    │               new-2.txt               │
                                             │     B/op     │     B/op      vs base                 │
UnmarshalPayload/UnmarshalPayload-12           6.974Ki ± 0%   6.974Ki ± 0%       ~ (p=1.000 n=10) ¹
UnmarshalPayload/UnmarshalPayloadComplete-12     997.0 ± 0%     997.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                        2.606Ki        2.606Ki       +0.00%
¹ all samples are equal

                                             │  new.txt   │              new-2.txt              │
                                             │ allocs/op  │ allocs/op   vs base                 │
UnmarshalPayload/UnmarshalPayload-12           13.00 ± 0%   13.00 ± 0%       ~ (p=1.000 n=10) ¹
UnmarshalPayload/UnmarshalPayloadComplete-12   12.00 ± 0%   12.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                        12.49        12.49       +0.00%
¹ all samples are equal
```